+++ examples/java-server/src/main/java/com/twitter/hello/app/HelloWorldApp.java
@@ -5,16 +5,16 @@ import java.util.Collection;
 import com.google.common.collect.ImmutableList;
 import com.google.inject.Module;
 
-import com.twitter.finatra.logging.modules.Slf4jBridgeModule$;
 import com.twitter.hello.HelloService;
 import com.twitter.inject.app.AbstractApp;
+import com.twitter.inject.logging.modules.LoggerModule$;
 
 public class HelloWorldApp extends AbstractApp {
 
     @Override
     public Collection<Module> javaModules() {
         return ImmutableList.<Module>of(
-                Slf4jBridgeModule$.MODULE$);
+                LoggerModule$.MODULE$);
     }
 
     @Override