+++ inject/inject-slf4j/src/main/scala/com/twitter/inject/logging/Slf4jBridgeUtility.scala
@@ -1,13 +1,16 @@
-package com.twitter.finatra.logging.modules
+package com.twitter.inject.logging
 
-import com.twitter.inject.{Injector, TwitterModule}
+import com.twitter.inject.Logging
+import org.slf4j.LoggerFactory
 import org.slf4j.bridge.SLF4JBridgeHandler
 
-object Slf4jBridgeModule extends TwitterModule {
-  override def singletonStartup(injector: Injector) {
+object Slf4jBridgeUtility extends Logging {
+
+  private[inject] def attemptSlf4jBridgeHandlerInstallation(): Unit = {
     if (canInstallBridgeHandler) {
       SLF4JBridgeHandler.removeHandlersForRootLogger()
       SLF4JBridgeHandler.install()
+      info("org.slf4j.bridge.SLF4JBridgeHandler installed.")
     }
   }
 
@@ -18,6 +21,7 @@ object Slf4jBridgeModule extends TwitterModule {
     // exists on the classpath. See: http://www.slf4j.org/legacy.html#jul-to-slf4j
     try {
       Class.forName("org.slf4j.impl.JDK14LoggerFactory", false, this.getClass.getClassLoader)
+      LoggerFactory.getLogger(this.getClass).warn("Detected [org.slf4j.impl.JDK14LoggerFactory] on classpath. SLF4JBridgeHandler cannot be installed, see: http://www.slf4j.org/legacy.html#jul-to-slf4j")
       false
     } catch {
       case e: ClassNotFoundException =>